+++ robotframework-ls/docs/changelog.md
@@ -15,6 +15,8 @@ NEXT
 - Properly recognize variable that evaluates to 0.
 - Properly recognize variables from Python that start with `DICT__` or `LIST__`.
 - Fix syntax highlight of inline python evaluation inside of variables in IF statements. [#666](https://github.com/robocorp/robotframework-lsp/issues/666) 
+- Properly consider that the `Comment` keyword arguments should be handled as comments. [#665](https://github.com/robocorp/robotframework-lsp/issues/665)
+- Created `robot.lint.ignoreEnvironmentVariables` setting so that it's possible to ignore environment variables not found. [#663](https://github.com/robocorp/robotframework-lsp/issues/663)
 
 
 New in 0.46.0 (2022-04-19)